requirements.txt
cklass
pyyaml
django
djangorestframework
gunicorn
whitenoise
dj-database-url
django-annoying
